\name{condition}
\alias{condition}
\docType{data}
\title{
The conditions of Mont&pick data
}
\description{
The conditions of Montgomery and Pickrell data. There are 60 CEU and 69 YRI.
}
\usage{data("condition")}

\details{
See count.data help page.
}
\source{
  https://bioconductor.org/packages/release/bioc/html/tweeDEseq.html
}
\references{
  Juan R Gonzalez and Mikel Esnaola (2016). tweeDEseqCountData: RNA-seq count data
  employed in the vignette of the tweeDEseq package. R package version 1.12.0.

}
\examples{
data(condition)

}
\seealso{
\code{\link{count.data}}
}
\keyword{datasets}
