\name{intSEQ-package}
\alias{intSEQ-package}

\docType{package}
\title{
Differential Analysis of RNA-seq Data with Integrated Likelihood Method
}
\description{
Differential expression analysis of RNA-seq data using integrated likelihood method. We fit negative binomial model and use integrated likelihood ratio test.

}


\author{
Yilun Zhang <yilzhang@ucdavis.edu>

}
\references{
The integrated paper}
