\name{intSEQ}
\alias{intSEQ}

\alias{intSEQ.default}
\alias{intSEQ.DGEList}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Implement integrated likelihood ratio test.
}
\description{
intSEQ implement integrated likelihood ratio test, returns a vector of p-values corresponds to rows of count.data.
}
\usage{
\S3method{intSEQ}{default}(count.data, condition,  nullcondition = NULL,nneighbour = 400
, lambda1 =ncol(count.data),lambda2 =0.05, 
meanmeth = c("estimator", "local.mean"), smoothmethod = c("loess", "spline", "no"), 
normalize = TRUE,  offsets = NULL, weights = NULL,   constadj=FALSE, ...)

\S3method{intSEQ}{DGEList}(count.data,  nullcondition = NULL,nneighbour = 400,
lambda1 =ncol(count.data),lambda2 =0.05,
meanmeth = c("estimator", "local.mean"), smoothmethod = c("loess", "spline", "no"),
normalize = TRUE, offsets = NULL, weights = NULL,  constadj=FALSE, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{count.data}{
The count table of RNA-seq data, should be a numerical matrix, or a DGEList object generated by DGEList function in edgeR.
}

  \item{condition}{
The conditions of RNA-seq data of the full model, should be a vector with length equal to number of columns of count.data or a data frame with number of rows equal to number of columns of count.data.
}

\item{nullcondition}{The conditions of RNA-seq data of the full model, should be a vector with length equal to number of columns of count.data or a data frame with number of rows equal to number of columns of count.data. The default value is NULL, means the global null model.

}

  \item{nneighbour}{
Number of neighbours selected to estimate the mean and variance of the normal prior.
}
  \item{lambda1}{
The first parameter of the variance, see details below.
}
  \item{lambda2}{
The second parameter of the variance, see details below.
}
  \item{meanmeth}{
Use the estimated dispersion or the local mean as the mean of normal prior.
}
  \item{smoothmethod}{
The smooth method to get the mean-dispersion trend.
}
  \item{normalize}{
Whether use "TMM" method to normalize the count data.
}


  \item{offsets}{
 Give offsets for the log-linear models. Should be a vector with length of ncol(count.data)
}

  \item{weights}{
optional numeric matrix giving prior weights for the observations (for each library and gene) to be used in the GLM calculations.
}
\item{constadj}{
Whether a constant should be multiplied to adjust the underflow problem of joint distribution. 
}
  \item{...}{
Other arguments that are currently not used.
}
}
\details{
intSEQ implements the integrated likelihood ratio test of RNA-seq data developed by (the new paper).

We first use estimateGLMTrendedDisp and estimateGLMTagwiseDisp in edgeR to estimate the Cox and Reid's dispersion. A mean-dispersion trend was estimated either with leoss or spline. 

For each gene, find 100 neighbors of that gene w.r.t the mean. Estimate he local mean of dispersion by the smoothing fit. Estimate the sample standard deviation by \eqn{s=\max \{ \hat{\sigma},0.05 \} }. \eqn{\hat{\sigma}}{hat{sigma}} is the square root of MSE of smoothing fit.

Let the prior be \eqn{\pi(\theta)=N(\bar{\theta},\lambda_1 s^2+\lambda_2)}
 
 Calculate the integrated likelihood for the null and alternative 
 		\eqn{		L_0=	\int L(Y | \mu_{0}, \theta) \pi(\theta) d\theta
 				} and  \eqn{
 				L_a=	\int L(Y | \mu_{a}, \theta) \pi(\theta) d\theta
 				}. \eqn{\mu_0} and \eqn{\mu_a} are the MLE of mean under null and alternative.
 Then compare \eqn{-2(\log L_0-\log L_a)} with chi-square distribution.
}
\value{
An object of class "intres". A list containing 3 object. The first "restable" stores most information, the last two for internal use only.
"restable" contains following
\item{logFC}{log fold change.}
\item{logCPM}{the log-average concentration/abundance for each tag in the two groups being compared.}
\item{FDR}{The FDR adjusted p values of integrated likelihood.}
\item{intLR}{The integrated likelihood ratio statistics.}
\item{intPValue}{The integrated likelihood ratio p values.}
\item{ordinaryLR}{The ordinary likelihood ratio statistics.}
\item{ordinaryPValue}{The ordinary likelihood ratio p values.}
The "parameters" contains fitted values of the count table and estimated dispersion parameters.
The "cond" stores The conditions of RNA-seq data.

}
\references{
our paper
}
\author{
Yilun Zhang, David M. Rocke
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
#select the first 10 columns of mont&pick data
data(count.data)
data(condition)
count=count.data[,1:10]
cond=rep(0:1,each=5)
res=intSEQ(count, cond)
res$restable[1:20,]

#Do the same thing with DGEList object
library(edgeR)
dge=DGEList(counts=count, group=cond)
res2=intSEQ(dge)
res2$restable[1:20,]
}

